reorganize in prep for full validation rules
