+++ back/zomes/projects/src/project/edge.rs
@@ -1,3 +1,3 @@
-pub mod edge;
+pub mod crud;
 pub mod error;
 pub mod validate;
