+++ back/zomes/projects/src/project/edge/crud.rs
@@ -38,11 +38,3 @@ impl Edge {
         }
     }
 }
-
-#[cfg(test)]
-mod tests {
-    #[test]
-    fn it_works() {
-        assert_eq!(2 + 2, 4);
-    }
-}
